+++ bindings/nodejs/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## \[0.3.7]
+
+- Fixes updating client options.
+  - [8124c3de](https://github.com/iotaledger/wallet.rs/commit/8124c3de93f006c20a1e0640d89fbfb6ab42d325) use iota_client ([#611](https://github.com/iotaledger/wallet.rs/pull/611)) on 2021-05-10
+
+## \[0.3.6]
+
+- Fix rev in `native/Cargo.toml`
+  - [c2caecaa](https://github.com/iotaledger/wallet.rs/commit/c2caecaaa69ad5fd9a98c346f3c3a599573679cc) fix(ci): Fix covector config on 2021-05-05
+
+## \[0.3.5]
+
+- Set git repo and rev to allow JS bindings to be built from source
+  - [ab7556e8](https://github.com/iotaledger/wallet.rs/commit/ab7556e88322e89fb37876186a5bca1329c752fc) fix(bindings/nodejs): Set rev to allow building from source ([#601](https://github.com/iotaledger/wallet.rs/pull/601)) on 2021-05-05
+
 ## \[0.3.4]
 
 - Fix `rocksdb` publish issue on v0.3.3